+++ javascriptv3/example_code/dynamodb/actions/batch-get-item.js
@@ -21,7 +21,7 @@ node ddb_batchgetitem.js
 // snippet-start:[dynamodb.JavaScript.batch.GetItemV3]
 // Import required AWS SDK clients and commands for Node.js
 import { BatchGetItemCommand } from "@aws-sdk/client-dynamodb";
-import { ddbClient } from "./libs/ddbClient.js";
+import { ddbClient } from "../src/libs/ddbClient.js";
 
 // Set the parameters
 export const params = {
