+++ javascriptv3/example_code/dynamodb/actions/batch-write-item.js
@@ -21,7 +21,7 @@ node ddb_batchwriteitem.js
 // snippet-start:[dynamodb.JavaScript.batch.WriteItemV3]
 // Import required AWS SDK clients and commands for Node.js
 import { BatchWriteItemCommand } from "@aws-sdk/client-dynamodb";
-import { ddbClient } from "./libs/ddbClient.js";
+import { ddbClient } from "../src/libs/ddbClient.js";
 
 // Set the parameters
 export const params = {
