+++ javascriptv3/example_code/dynamodb/actions/describe-table.js
@@ -6,34 +6,32 @@ which is available at https://github.com/aws/aws-sdk-js-v3. This example is in t
 https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/dynamodb-examples-using-tables.html.
 
 Purpose:
-ddb_deletetable.js demonstrates how to delete an Amazon DynamoDB table.
+ddb_describetable.js demonstrates how to retrieve information about an Amazon DynamoDB table.
 
 INPUTS:
 - TABLE_NAME
 
 Running the code:
-node ddb_deletetable.js
-
+node ddb_describetable.js
 */
-// snippet-start:[dynamodb.JavaScript.item.deleteTableV3]
+// snippet-start:[dynamodb.JavaScript.table.describeTableV3]
 // Import required AWS SDK clients and commands for Node.js
-import { DeleteTableCommand } from "@aws-sdk/client-dynamodb";
-import { ddbClient } from "./libs/ddbClient.js";
+import { DescribeTableCommand } from "@aws-sdk/client-dynamodb";
+import { ddbClient } from "../src/libs/ddbClient.js";
 
 // Set the parameters
-export const params = {
-  TableName: "CUSTOMER_LIST_NEW",
-};
+export const params = { TableName: "TABLE_NAME" }; //TABLE_NAME
 
 export const run = async () => {
   try {
-    const data = await ddbClient.send(new DeleteTableCommand(params));
-    console.log("Success, table deleted", data);
+    const data = await ddbClient.send(new DescribeTableCommand(params));
+    console.log("Success", data);
+    // console.log("Success", data.Table.KeySchema);
     return data;
   } catch (err) {
-      console.log("Error", err);
-    }
+    console.log("Error", err);
+  }
 };
 run();
-// snippet-end:[dynamodb.JavaScript.item.deleteTableV3]
+// snippet-end:[dynamodb.JavaScript.table.describeTableV3]
