+++ javascriptv3/example_code/dynamodb/actions/list-tables.js
@@ -15,7 +15,7 @@ node ddb_listtables.js
 
 // Import required AWS SDK clients and commands for Node.js
 import { ListTablesCommand } from "@aws-sdk/client-dynamodb";
-import { ddbClient } from "./libs/ddbClient.js";
+import { ddbClient } from "../src/libs/ddbClient.js";
 
 export const run = async () => {
   try {
