+++ javascriptv3/example_code/dynamodb/actions/put-item.js
@@ -18,7 +18,7 @@ node ddb_putitem.js
 
 // Import required AWS SDK clients and commands for Node.js
 import { PutItemCommand } from "@aws-sdk/client-dynamodb";
-import { ddbClient } from "./libs/ddbClient.js";
+import { ddbClient } from "../src/libs/ddbClient.js";
 
 // Set the parameters
 export const params = {
